Moved Auth related items to token-verifier-util
